Refactor para sacar las constantes del codigo y separar las funciones del servicio de WPGQL de las Queries
