+++ pages/blog/categorias/[categorySlug].js
@@ -12,7 +12,7 @@ import { ApolloProvider, useLazyQuery } from '@apollo/client';
 import { apolloClient } from "../../../config/apollo";
 import RelatedPostsSection from '../../../components/Blog/RelatedPostsSection';
 import { ObjectInspector, chromeDark } from 'react-inspector';
-import { QUERY_GET_POSTS_BY_CATEGORY_ID } from "../../../services/wordpressGQL";
+import { QUERY_GET_POSTS_BY_CATEGORY_ID } from "../../../services/wordpressGQLQueries";
 import { BLOG_LOCALES, BLOG_RENDER_CATEGORIES } from '../../../constants/blog';
